--- packages/css/src/breadcrumb/breadcrumb-css.md
@@ -1,1 +1,0 @@
-# CSS Breadcrumb
