+++ packages/react/src/Breadcrumb/Breadcrumb.tsx
@@ -1,11 +1,5 @@
-import React, { forwardRef, useMemo } from 'react'
-import type { ForwardedRef, FunctionComponent, HTMLAttributes, PropsWithChildren } from 'react'
-
-function isReactFunctionCom<Tprops>(
-  element: React.ReactNode,
-): element is React.ReactElement<Tprops, FunctionComponent> {
-  return React.isValidElement(element)
-}
+import React, { forwardRef } from 'react'
+import type { ForwardedRef, HTMLAttributes, PropsWithChildren } from 'react'
 
 interface BreadcrumbComponent
   extends React.ForwardRefExoticComponent<
@@ -14,26 +8,11 @@ interface BreadcrumbComponent
   Item: typeof BreadcrumbItem
 }
 
-// We're validating against an index. Maxiumum children are 8.
-const breadcrumbLimit = 8
-
 export const Breadcrumb = forwardRef(
   ({ children, ...restProps }: PropsWithChildren<HTMLAttributes<HTMLElement>>, ref: ForwardedRef<HTMLElement>) => {
-    const allowedBreadcrumbItems = useMemo(() => {
-      return React.Children.toArray(children).filter((child, index) => {
-        if (isReactFunctionCom(child) && child.type.displayName !== 'BreadcrumbItem') {
-          // TODO: Do we want to log this as a warning?
-          console.warn(`Breadcrumb: ${child.type.displayName ?? child.type} is not a valid child`)
-          return false
-        }
-
-        return index < breadcrumbLimit
-      })
-    }, [children])
-
     return (
       <nav {...restProps} className="amsterdam-breadcrumb" ref={ref}>
-        <ol className="amsterdam-breadcrumb-list">{allowedBreadcrumbItems}</ol>
+        <ol className="amsterdam-breadcrumb__list">{children}</ol>
       </nav>
     )
   },
@@ -41,20 +20,21 @@ export const Breadcrumb = forwardRef(
 
 Breadcrumb.displayName = 'Breadcrumb'
 
-interface BreadcrumbItemProps extends PropsWithChildren {
+export interface BreadcrumbItemProps extends PropsWithChildren<HTMLAttributes<HTMLLIElement>> {
   href: string
 }
 
-// TODO: We might wanna wrap this in a forwardRef as well.
-const BreadcrumbItem = ({ children, href }: BreadcrumbItemProps) => {
-  return (
-    <li className="amsterdam-breadcrumb-item">
-      <a className="amsterdam-breadcrumb-link" href={href}>
-        {children}
-      </a>
-    </li>
-  )
-}
+const BreadcrumbItem = forwardRef(
+  ({ children, href, ...restProps }: BreadcrumbItemProps, ref: ForwardedRef<HTMLLIElement>) => {
+    return (
+      <li {...restProps} className="amsterdam-breadcrumb-item" ref={ref}>
+        <a className="amsterdam-breadcrumb-item__link" href={href}>
+          {children}
+        </a>
+      </li>
+    )
+  },
+)
 
 BreadcrumbItem.displayName = 'BreadcrumbItem'
 Breadcrumb.Item = BreadcrumbItem
