+++ packages/react/src/Breadcrumb/index.ts
@@ -1,1 +1,2 @@
 export { Breadcrumb } from './Breadcrumb'
+export type { BreadcrumbItemProps } from './Breadcrumb'
