+++ storybook/storybook-css/src/components/Breadcrumb/Breadcrumb.docs.mdx
@@ -1,12 +1,9 @@
 import { Markdown, Meta, Primary } from "@storybook/blocks";
 import * as Breadcrumb from "./Breadcrumb.stories.tsx";
 import README from "../../../../../packages/css/src/breadcrumb/README.md?raw";
-import css from "../../../../../packages/css/src/breadcrumb/breadcrumb-css.md?raw";
 
 <Meta of={Breadcrumb} />
 
 <Markdown>{README}</Markdown>
-
-<Markdown>{css}</Markdown>
 
 <Primary />
